chore: Refactor padding in Header and StairsHeader menu styles
The code changes in this commit refactor the padding values in the Header and StairsHeader menu styles. The padding in the Header component is changed to allow the text to stick to the edges in the X-axis. In the StairsHeader menu component, the padding is replaced with the padding-block property to ensure consistent spacing. These changes improve the layout and visual appearance of the header components.